fix bug

+++ src/edt_utils.py
@@ -287,7 +287,7 @@ def process_line(line_number, labeled_line, offset, line_leads, config_dict):
     for label in segment_labels:
 
         roi = (labeled_line == label)
-        sl = ndimage.find_objects(roi)
+        sl = ndimage.find_objects(roi.astype(np.int32))
 
         if len(sl) == 0:
             continue
